+++ qcetl/column.py
@@ -1786,7 +1786,7 @@ class SequenzaColumn(BaseColumn):
 
 class UltimaLibraryMetricsColumn(BaseColumn):
     Run = ColumnNames.Run
-    SampleName = "Sample"
+    UltimaLibraryID = "Ultima Library ID"
     Barcode = "Barcode"
     PineryLimsID = ColumnNames.PineryLimsID
     MeanCoverage = "mean coverage"